requirements.txt
PyPDF2
ebooklib
PyYAML
tqdm
langchain
